refactor: split daejeon event sync fetch and normalize
