+++ scripts/daejeon-event-sync/constants.ts
@@ -1,0 +1,3 @@
+export const SEARCH_URL = 'https://www.daejeon.go.kr/fvu/FvuEventList.do?menuSeq=504';
+export const DETAIL_BASE_URL = 'https://www.daejeon.go.kr/fvu/FvuEventView.do';
+export const SOURCE_NAME = 'Daejeon Official Event Search';
